random: add check task and update Golang build image
Add a new task for the random application called "check".
It runs golint.

The commits also :
- updates the Golang docker image used during build and check,
- removes passing -mod=vendor to "go build", using vendored libs if vendor/
  exist is now the default for "go build"
- updates the vendor directory

+++ b-team/random/build.sh
@@ -2,10 +2,23 @@
 
 set -eux -o pipefail
 
-DOCKER_IMG="golang@sha256:e7462ca504afc789d289f2bb5fd471815cc11833439d2fe4e61915b190045359"
+DOCKER_IMG="golang@sha256:2de21f4fe5c3557c31b716291a7986c0d54abbd282cd7fb43aaa484672b0c855"
 SCRIPT_DIR="$(cd "$(dirname "$0")" ; pwd -P)"
 
-docker run --rm --init -e HOME=/tmp/ --user="$(id -u):$(id -g)" \
-           -v "$SCRIPT_DIR":/code -w /code  \
-	   $DOCKER_IMG \
-           go build -mod=vendor -o "dist/randomnumber" main.go
+if [ $# -ne 1 ]; then
+	echo "usage: $0 check|build"
+	exit 1
+fi
+
+if [ "$1" = "build" ]; then
+	docker run --rm --init -e HOME=/tmp/ --user="$(id -u):$(id -g)" \
+		   -v "$SCRIPT_DIR":/code -w /code  \
+		   $DOCKER_IMG \
+		   go build -o "dist/randomnumber" main.go
+elif [ "$1" = "check" ]; then
+	docker run --rm --init -e HOME=/tmp/ --user="$(id -u):$(id -g)" \
+		   -v "$SCRIPT_DIR":/code -w /code  \
+		   $DOCKER_IMG \
+		   go vet ./...
+fi
+
